[TASK] Disable compression for inlineJs in Backend
JSMin is a major performance nightmare and no network connection can be
as slow as it would justify to spend so much time to execute JSmin
several times for each and every backend request.

As long as we have no fast replacement in core available, disable
compression for all JS snippets added with
t3lib_PageRenderer::addJsInlineCode in the Backend.

JSMin's share of execution time was about 20-25% of the page generation
time, for almost empty backend pages even ~50%.

Change-Id: I5623ad6c8b9d24777b6bcf46ea97bf51eb9c62d3
Resolves: #29498