[BUGFIX] meaningfulTempFilePrefix is not multi-byte safe
With meaningfulTempFilePrefix, a prefix for file names of generated
image files can be set. The unit is currently (obviously) bytes, not
characters.

This causes trouble with $TYPO3_CONF_VARS['SYS']['UTF8filesystem'], when
the string is truncated exactly in the middle of a multi-byte char,
resulting in the file not being accessible via HTTP.
The file name "abcÄ" with a prefix length of 4 would cut the last byte
of the Ä, resulting in "abc?".

Thus, substring generation has to multi-byte safe and use
t3lib_cs::substring.

Change-Id: I9035b62ae6558f03d5a37b222175109a9ef6be58
Resolves: #30416
Releases: 4.6, 4.5