[BUGFIX] Exception on modifying page properties
An exception "#1300096564: uid of file has to be numeric" is
thrown on modifying and saving the properties of a pages record.
It's required that there are zero file elements on that page,
thus the FAL filter gets an faulty array which is not sanitized
before trying to get an accordant file.

Change-Id: I236610cebe872bcd3a4e49ceec4b24a26b2ef63b
Fixes: #38114
Releases: 6.0
Reviewed-on: http://review.typo3.org/12128
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel