[TASK] Move buttons in user setting into advanced functions
The buttons "clear configuration" and "reset
temporary data" are advanced functions and should
therefore be moved to the advanced tab.

Additionally the buttons divert the attention, so that
it is very hard to notice the save button on the
top.

The patch moves the buttons to the advanced tab and
additionally aligns the select box(es) with the
input and button fields.

Change-Id: I5c5552d3a20b68e458b6523bfb5e939000cd4f9a
Related: #632
Releases: 6.0
Resolves: #38523
Reviewed-on: http://review.typo3.org/12504
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn