[BUG] FAL AdapterService fails with Fluid
FAL introduced an adapterservice to migrate certain
fields to FAL style. The adapterservice uses type hinting
to suggest $row of tslib_content always to be an array.
However, when e.g. calling tslib_content from an
extension or Fluid, it is common to have $row
as NULL or empty, so the additional check
is removed.

Change-Id: I4b5f0231d3c8c2d9f3ec6d2f6a7a3a1b85775673
Resolves: #38977
Reviewed-on: http://review.typo3.org/12813
Reviewed-by: Michael Staatz
Tested-by: Michael Staatz
Reviewed-by: Ingmar Schlecht
Tested-by: Ingmar Schlecht