[BUGFIX] Remove debug statement for not existing classes
Due to the removal of allowed class name prefix check, 
all calls to t3lib_div::getUserObject cause a 
debug statement for not found class in BE while editing content. 
The debug statement is therefore removed.

Although silent failing is not nice, there is no other s
olution for now. In the long run we should deprecate 
the whole method in several steps.

Change-Id: I407d5ee6020913b6c0538d1414ccb6514d72c71b
Fixes: #39191
Releases: 6.0
Reviewed-on: http://review.typo3.org/13001
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel
Reviewed-by: Oliver Klee
Reviewed-by: Sebastian Michaelsen
Tested-by: Sebastian Michaelsen
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn